+++ aikau/src/test/resources/testApp/WEB-INF/classes/alfresco/site-webscripts/alfresco/documentlibrary/InfiniteScrollDocumentList.get.js
@@ -44,6 +44,9 @@ model.jsonModel = {
          }
       },
       {
+         name: "alfresco/html/HR"
+      },
+      {
          id: "DOCUMENT_LIST",
          name: "alfresco/documentlibrary/AlfDocumentList",
          config: {
@@ -73,15 +76,19 @@ model.jsonModel = {
 
 if (includeSortMenu)
 {
-   model.jsonModel.widgets.splice(1, 0, {
+   model.jsonModel.widgets.splice(2, 0, {
       name: "alfresco/menus/AlfMenuBar",
       config: {
          widgets: [
             {
+               name: "alfresco/documentlibrary/AlfSelectDocumentListItems"
+            },
+            {
                id: "DOCLIB_SORT_FIELD_SELECT",
                name: "alfresco/menus/AlfMenuBarSelect",
                config: {
-                  selectionTopic: "ALF_DOCLIST_SORT_FIELD_SELECTION",
+                  title: "Sort By",
+                  selectionTopic: "ALF_DOCLIST_SORT",
                   widgets: [
                      {
                         id: "DOCLIB_SORT_FIELD_SELECT_GROUP",
@@ -91,6 +98,34 @@ if (includeSortMenu)
                         }
                      }
                   ]
+               }
+            },
+            {
+               id: "SORT_TOGGLE",
+               name: "alfresco/menus/AlfMenuBarToggle",
+               config: {
+                  subscriptionTopic: "ALF_DOCLIST_SORT",
+                  subscriptionAttribute: "direction",
+                  checked: true,
+                  checkedValue: "ascending",
+                  onConfig: {
+                     title: "Change sort order to descending",
+                     iconClass: "alf-sort-ascending-icon",
+                     iconAltText: "Sorted ascending",
+                     publishTopic: "ALF_DOCLIST_SORT",
+                     publishPayload: {
+                        direction: "ascending"
+                     }
+                  },
+                  offConfig: {
+                     title: "Change sort order to ascending",
+                     iconClass: "alf-sort-descending-icon",
+                     iconAltText: "Sorted descending",
+                     publishTopic: "ALF_DOCLIST_SORT",
+                     publishPayload: {
+                        direction: "descending"
+                     }
+                  }
                }
             }
          ]